+++ test/test_helpers.py
@@ -1,23 +1,17 @@
 from __future__ import absolute_import, division, print_function, unicode_literals
 
-import random
 import unittest
 from math import degrees, radians, sqrt
 
 import numpy as np
-from six.moves import range
+from six.moves import range, zip
 
-from timezonefinder.global_settings import DTYPE_FORMAT_SIGNED_I_NUMPY, DTYPE_FORMAT_B_NUMPY, DTYPE_FORMAT_H_NUMPY, \
-    DTYPE_FORMAT_F_NUMPY, DECIMAL_PLACES_ACCURACY, COORD2INT_FACTOR, MAX_ALLOWED_COORD_VAL, INT2COORD_FACTOR
-
-
-def random_point():
-    # tzwhere does not work for points with more latitude!
-    return random.uniform(-180, 180), random.uniform(-84, 84)
-
-
-def list_of_random_points(length):
-    return [random_point() for i in range(length)]
+from auxiliaries import proto_test_case, random_point
+# from .auxiliaries import random_point, proto_test_case
+from timezonefinder.global_settings import (
+    COORD2INT_FACTOR, DECIMAL_PLACES_ACCURACY, DTYPE_FORMAT_F_NUMPY, DTYPE_FORMAT_H_NUMPY, DTYPE_FORMAT_SIGNED_I_NUMPY,
+    INT2COORD_FACTOR, MAX_ALLOWED_COORD_VAL,
+)
 
 
 def poly_conversion_fct(coords):
@@ -27,12 +21,6 @@ def poly_conversion_fct(coords):
     array = np.ndarray.astype(array, dtype=DTYPE_FORMAT_SIGNED_I_NUMPY)
     return array
 
-
-# from timezonefinder.helpers import (
-#     all_the_same, coord2int, distance_to_point_on_equator, distance_to_polygon, distance_to_polygon_exact,
-#     haversine,
-#     inside_polygon,
-# )
 
 class HelperTest(unittest.TestCase):
     import timezonefinder.helpers as helpers
@@ -47,8 +35,47 @@ class HelperTest(unittest.TestCase):
         "inside_polygon": helpers.inside_polygon,
         "coord2shortcut": helpers.coord2shortcut,
         "rectify_coordinates": helpers.rectify_coordinates,
+        'convert2coords': helpers.convert2coords,
+        'convert2coord_pairs': helpers.convert2coord_pairs,
     }
     print('\ntesting helpers.py functions...')
+
+    def test_coord2shortcut(self):
+
+        coord2shortcut = self.fct_dict['coord2shortcut']
+        if coord2shortcut is None:
+            print('test coord2shortcut skipped.')
+            return
+
+        def coord2shortcut_test_fct(input):
+            (lng, lat) = input
+            return coord2shortcut(lng, lat)
+
+        data = [
+            # shortcut numbering starts at "the top left" with x,y= 0,0
+            # always (only) the "top" and "left" borders belong to a shortcut
+            # the other borders belong to the next neighbouring shortcut
+            ((-180.0, 90.0), (0, 0)),
+            # shortcuts are constant for every 1 degree lng and 0.5 degree lat
+            # defined with NR_SHORTCUTS_PER_LNG, NR_SHORTCUTS_PER_LAT in timezonefinder.file_converter
+            ((-180.0 + INT2COORD_FACTOR, 90.0 - INT2COORD_FACTOR), (0, 0)),
+
+            # shortcut numbering follows the lng, lat coordinate grid
+            ((-179.0, 90.0), (1, 0)),
+            ((-178.9, 89.9), (1, 0)),
+            ((-180.0, 89.5), (0, 1)),
+            ((-179.9, 89.4), (0, 1)),
+            ((-180.0, 89.0), (0, 2)),
+            ((-179.9, 88.9), (0, 2)),
+
+            # shortcut numbering end at "the top left" with x,y= 359, 359
+            # lng= 180.0 == -180.0
+            # lat =-90.0 is not allowed (=bottom border of a shortcut)
+            ((180.0 - INT2COORD_FACTOR, -90 + INT2COORD_FACTOR), (359, 359)),
+            ((179.8, -89.8), (359, 359)),
+        ]
+
+        proto_test_case(data, coord2shortcut_test_fct)
 
     def test_dtype_conversion(self):
         coord2int = self.fct_dict['coord2int']
@@ -69,6 +96,30 @@ class HelperTest(unittest.TestCase):
         # backwards: int to coord
         values_converted_coords = [int2coord(x) for x in int_values]
         assert (np.all(np.array(values_converted_coords) == np.array(coord_values)))
+
+    def test_convert2coord_pairs(self):
+
+        convert2coord_pairs = self.fct_dict['convert2coord_pairs']
+        if convert2coord_pairs is None:
+            print('test convert2coord_pairs skipped.')
+            return
+
+        coord_values = [float(x) for x in range(-2, 3, 1)]
+        coord_polygon = (coord_values, coord_values)
+        polygon_int = poly_conversion_fct(coord_polygon)
+        assert (convert2coord_pairs(polygon_int) == list(zip(coord_values, coord_values)))
+
+    def test_convert2coords(self):
+
+        convert2coords = self.fct_dict['convert2coords']
+        if convert2coords is None:
+            print('test convert2coord_pairs skipped.')
+            return
+
+        coord_values = [float(x) for x in range(-2, 3, 1)]
+        coord_polygon = (coord_values, coord_values)
+        polygon_int = poly_conversion_fct(coord_polygon)
+        assert (convert2coords(polygon_int) == [coord_values, coord_values])
 
     # use only numpy data structures, because the functions are reused for testing the numba helpers
     def test_inside_polygon(self):
@@ -347,8 +398,9 @@ class HelperTestNumba(HelperTest):
             "inside_polygon": helpers.inside_polygon,
             "coord2shortcut": helpers.coord2shortcut,
             "rectify_coordinates": helpers.rectify_coordinates,
+            'convert2coords': helpers.convert2coords,
+            'convert2coord_pairs': helpers.convert2coord_pairs,
         }
-
         print('\nNumba installation found.\ntesting helpers_numba.py functions...')
 
     except ImportError:
@@ -363,6 +415,8 @@ class HelperTestNumba(HelperTest):
             "inside_polygon": None,
             "coord2shortcut": None,
             "rectify_coordinates": None,
+            'convert2coords': None,
+            'convert2coord_pairs': None,
         }
         print('\nNumba installation NOT found.')