+++ test/test_it.py
@@ -1,27 +1,15 @@
 from __future__ import absolute_import, division, print_function, unicode_literals
 
-import random
 import unittest
 from datetime import datetime
 
 import pytest
-from six.moves import range
 
-from timezonefinder.timezonefinder import TimezoneFinder, convert2coord_pairs, convert2coords, coord2shortcut
+from auxiliaries import list_of_random_points, random_point
+from timezonefinder.global_settings import INT2COORD_FACTOR
+from timezonefinder.timezonefinder import TimezoneFinder
 
-
-def proto_test_case(data, fct):
-    for input, expected_output in data:
-        # print(input, expected_output, fct(input))
-        actual_output = fct(input)
-        if actual_output != expected_output:
-            print('input: {} expected: {} got: {}'.format(input, expected_output, actual_output))
-        assert actual_output == expected_output
-
-
-# removed tzwhere comparison, since it is outdated
-# from tzwhere.tzwhere import tzwhere
-# SHAPELY = True  # tzwhere requires shapely now
+# from .auxiliaries import random_point, list_of_random_points
 
 
 # number of points to test (in each test, realistic and random ones)
@@ -29,7 +17,6 @@ N = 100
 
 # mistakes in these zones don't count as mistakes
 excluded_zones_timezonefinder = []
-# excluded_zones_tzwhere = []
 
 TEST_LOCATIONS = [
     # lat, lng, description, correct output
@@ -97,7 +84,6 @@ TEST_LOCATIONS_CERTAIN = TEST_LOCATIONS + [
     # (55.6056074, 12.7128568, 'Oresund Bridge2',  None),
 ]
 
-
 TEST_LOCATIONS_PROXIMITY = [
     # the polygons in the new data do not follow the coastlines any more
     # proximity tests are not meaningful at the moment
@@ -113,74 +99,36 @@ TEST_LOCATIONS_PROXIMITY = [
 ]
 
 
-def random_point():
-    # tzwhere does not work for points with higher latitude!
-    return random.uniform(-180, 180), random.uniform(-84, 84)
+def print_time(timezoefinder_time):
+    print('required time:', timezoefinder_time)
 
 
-def list_of_random_points(length):
-    return [random_point() for i in range(length)]
-
-
-def print_speedup(timezoefinder_time):
-    # print('tzwhere:', tz_where_time)
-    print('timezonefinder:', timezoefinder_time)
-    # try:
-    #     speedup = round(tz_where_time / timezoefinder_time - 1, 2)
-    #     if speedup < 0:
-    #         print(round(timezoefinder_time / tz_where_time - 1, 2), 'times slower')
-    #     else:
-    #         print(speedup, 'times faster')
-    # except TypeError:
-    #     pass
-
-
-class PackageEqualityTest(unittest.TestCase):
+class MainPackageTest(unittest.TestCase):
     # do the preparations which have to be made only once
 
     print("startup time:")
-
-    # if SHAPELY:
-    #     print('shapely: ON (tzwhere)')
-    # else:
-    #     print('shapely: OFF (tzwhere)')
-
     if TimezoneFinder.using_numba():
-        print('Numba: ON (timezonefinder)')
+        print('Numba: ON (precompiled functions in use)')
     else:
-        print('Numba: OFF (timezonefinder)')
+        print('Numba: OFF (precompiled functions NOT in use)')
 
     start_time = datetime.now()
     timezone_finder = TimezoneFinder()
     end_time = datetime.now()
     my_time = end_time - start_time
-
-    # print('Starting tz_where. This could take a while...')
-
-    # integrated start up time test:
-    # (when doing this for multiple times things are already cached and therefore produce misleading results)
-    # start_time = datetime.now()
-    # tz_where = tzwhere()
-    # end_time = datetime.now()
-    # his_time = end_time - start_time
-
-    print_speedup(timezoefinder_time=my_time)
+    print_time(timezoefinder_time=my_time)
     print('\n')
 
-    # create an array of n points where tzwhere finds something (realistic queries)
+    # create an array of points where timezone_finder finds something (realistic queries)
     print('collecting and storing', N, 'realistic points for the tests...')
     realistic_points = []
-    # real_ps_results_tzwhere = []
-    # real_ps_results_certain = []
-
     ps_for_10percent = int(N / 10)
     percent_done = 0
 
     i = 0
     while i < N:
         lng, lat = random_point()
-        # a realistic point is a point where certain_timezone_at() or tzwhere find something
-        # if not (tz_where.tzNameAt(lat, lng) or timezone_finder.certain_timezone_at(lng=lng, lat=lat)):
+        # a realistic point is a point where certain_timezone_at() finds something
         if timezone_finder.certain_timezone_at(lng=lng, lat=lat):
             i += 1
             realistic_points.append((lng, lat))
@@ -193,14 +141,7 @@ class PackageEqualityTest(unittest.TestCase):
     def test_speed(self):
         print("\n\nSpeed Tests:\n-------------")
 
-        # def check_speed_tzwhere(list_of_points):
-        #     start_time = datetime.now()
-        #     for point in list_of_points:
-        #         self.tz_where.tzNameAt(latitude=point[1], longitude=point[0])
-        #     end_time = datetime.now()
-        #     return end_time - start_time
-
-        def check_speed_my_algor(list_of_points):
+        def check_speed_of_algorithm(list_of_points):
             start_time = datetime.now()
             for point in list_of_points:
                 self.timezone_finder.timezone_at(lng=point[0], lat=point[1])
@@ -208,15 +149,9 @@ class PackageEqualityTest(unittest.TestCase):
             return end_time - start_time
 
         def print_speed_test(type_of_points, list_of_points):
-            my_time = check_speed_my_algor(list_of_points)
-            # his_time = check_speed_tzwhere(list_of_points)
-            print('\nTIMES for ', N, type_of_points)
-            print_speedup(timezoefinder_time=my_time)
-
-        # if SHAPELY:
-        #     print('shapely: ON (tzwhere)')
-        # else:
-        #     print('shapely: OFF (tzwhere)')
+            my_time = check_speed_of_algorithm(list_of_points)
+            print('\nrequired time for ', N, type_of_points)
+            print_time(timezoefinder_time=my_time)
 
         if TimezoneFinder.using_numba():
             print('Numba: ON (timezonefinder)')
@@ -233,14 +168,13 @@ class PackageEqualityTest(unittest.TestCase):
         assert self.timezone_finder.timezone_at(lng=-180.0, lat=-90.0) == 'Antarctica/McMurdo'
 
         with pytest.raises(ValueError):
-            self.timezone_finder.timezone_at(lng=180.01, lat=90.0)
-            self.timezone_finder.timezone_at(lng=180.0, lat=90.01)
-            self.timezone_finder.timezone_at(lng=-180.01, lat=90.0)
-            self.timezone_finder.timezone_at(lng=-180.0, lat=90.01)
-            self.timezone_finder.timezone_at(lng=180.01, lat=-90.0)
-            self.timezone_finder.timezone_at(lng=180.0, lat=-90.01)
-            self.timezone_finder.timezone_at(lng=-180.01, lat=-90.0)
-            self.timezone_finder.timezone_at(lng=-180.0, lat=-90.01)
+            self.timezone_finder.timezone_at(lng=180.0 + INT2COORD_FACTOR, lat=90.0)
+            self.timezone_finder.timezone_at(lng=-180.0 - INT2COORD_FACTOR, lat=90.0 + INT2COORD_FACTOR)
+            self.timezone_finder.timezone_at(lng=-180.0, lat=90.0 + INT2COORD_FACTOR)
+            self.timezone_finder.timezone_at(lng=180.0 + INT2COORD_FACTOR, lat=-90.0)
+            self.timezone_finder.timezone_at(lng=180.0, lat=-90.0 - INT2COORD_FACTOR)
+            self.timezone_finder.timezone_at(lng=-180.0 - INT2COORD_FACTOR, lat=-90.0)
+            self.timezone_finder.timezone_at(lng=-180.0 - INT2COORD_FACTOR, lat=-90.01 - INT2COORD_FACTOR)
 
     def test_kwargs_only(self):
         # calling timezonefinder fcts without keyword arguments should raise an error
@@ -295,81 +229,6 @@ class PackageEqualityTest(unittest.TestCase):
             print(template.format(loc, str(expected), str(computed), ok))
 
         assert no_mistakes_made
-
-    # disabled because underlying data is completely different atm
-    # def test_equality(self):
-    #     # Test the equality of the two packages for N realistic and N random points
-    #     def print_equality_test(types_of_points, list_of_points):
-    #         print('\ntesting', N, types_of_points)
-    #         print('MISMATCHES:')
-    #         template = '{0:40s} | {1:20s} | {2:21s} | {3:20s}'
-    #         print(template.format('Point', 'timezone_at()', 'certain_timezone_at()', 'tzwhere'))
-    #         print('=========================================================================')
-    #         mistakes = 0
-    #         for lng, lat in list_of_points:
-    #             his_result = self.tz_where.tzNameAt(lat, lng)
-    #             my_result_certain = self.timezone_finder.certain_timezone_at(lng=lng, lat=lat)
-    #             # test only makes sense if certain_timezone_at() or tzwhere find something
-    #             if his_result is not None or my_result_certain is not None:
-    #                 my_result = self.timezone_finder.timezone_at(lng=lng, lat=lat)
-    #                 if my_result != his_result or my_result_certain != his_result:
-    #                     if his_result in excluded_zones_tzwhere and my_result in excluded_zones_timezonefinder:
-    #                         print(template.format((lat, lng), my_result, my_result_certain,
-    #                                               his_result), '(not counted, see issue section)')
-    #                     else:
-    #                         mistakes += 1
-    #                         print(template.format(str((lat, lng)), str(my_result), str(my_result_certain),
-    #                                               str(his_result)))
-    #         print('\nin', N, 'tries', mistakes, 'mismatches were made')
-    #         fail_percentage = mistakes * 100 / (2 * N)
-    #         assert fail_percentage < 5
-    #
-    #     print_equality_test('realistic points', self.realistic_points)
-    #     print_equality_test('random points', list_of_random_points(length=N))
-
-    @staticmethod
-    def test_convert2coord_pairs():
-        data = [[10000000, 20000000, 30000000], [10000000, 20000000, 30000000]]
-        # print(convert2coord_pairs(data))
-        assert (convert2coord_pairs(data) == [(1.0, 1.0), (2.0, 2.0), (3.0, 3.0)])
-
-    @staticmethod
-    def test_convert2coords():
-        data = [[10000000, 20000000, 30000000], [10000000, 20000000, 30000000]]
-        assert (convert2coords(data) == [[1.0, 2.0, 3.0], [1.0, 2.0, 3.0]])
-
-    @staticmethod
-    def test_coord2shortcut():
-
-        def coord2shortcut_test_fct(input):
-            (lng, lat) = input
-            return coord2shortcut(lng, lat)
-
-        data = [
-            # shortcut numbering starts at "the top left" with x,y= 0,0
-            # always (only) the "top" and "left" borders belong to a shortcut
-            # the other borders belong to the next neighbouring shortcut
-            ((-180.0, 90.0), (0, 0)),
-            # shortcuts are constant for every 1 degree lng and 0.5 degree lat
-            # defined with NR_SHORTCUTS_PER_LNG, NR_SHORTCUTS_PER_LAT in timezonefinder.file_converter
-            ((-179.9, 89.9), (0, 0)),
-
-            # shortcut numbering follows the lng, lat coordinate grid
-            ((-179.0, 90.0), (1, 0)),
-            ((-178.9, 89.9), (1, 0)),
-            ((-180.0, 89.5), (0, 1)),
-            ((-179.9, 89.4), (0, 1)),
-            ((-180.0, 89.0), (0, 2)),
-            ((-179.9, 88.9), (0, 2)),
-
-            # shortcut numbering end at "the top left" with x,y= 359, 359
-            # lng= 180.0 == -180.0
-            # lat =-90.0 is not allowed (=bottom border of a shortcut)
-            ((179.9, -89.9), (359, 359)),
-            ((179.8, -89.8), (359, 359)),
-        ]
-
-        proto_test_case(data, coord2shortcut_test_fct)
 
     def test_overflow(self):
         longitude = -123.2